+++ lib/data_models/db_record.dart
@@ -6,9 +6,7 @@ abstract class DBRecord {
   @required
   final dynamic id;
 
-  DBRecord fromJson(Map<String, dynamic> r) {
-    return null;
-  }
+  DBRecord.fromJson(Map<String, dynamic> r) : id = r['id'];
 
   Map<String, dynamic> toJson() {
     return null;